Верблюд-быстроход

+++ RaceSimulation/GroundVehicle.cpp
@@ -15,7 +15,7 @@ int GroundVehicle::getRestNumber(const int distanceKm) const
 {
     const auto numberOfMotionPeriods = getMotionTimeHr(distanceKm) / _motionTimeBeforeRestHr;
     // если пришли на финиш в точности, когда надо отдыхать, то считаем, что отдых уже не нужен
-    return numberOfMotionPeriods < 1 ? 0 : std::ceil(numberOfMotionPeriods) - 1;
+    return numberOfMotionPeriods < 1 ? 0 : static_cast<int>(std::ceil(numberOfMotionPeriods) - 1);
 }
 
 double GroundVehicle::getRaceTimeHr(const int distanceKm) const
